requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
celery[redis]==5.0.5 
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
grpcio==1.35.0
grpcio-tools==1.35.0
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
msgpack==1.0.2
numpy==1.20.1
#opencv-python==4.5.4.58
opencv-python-headless==4.2.0.34
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3 
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
psycopg2-binary==2.8.6
wand==0.6.6
requests==2.25.1
django-cors-headers==3.7.0